1789: replace the GbDataTable class with inline overrides of just the methods we need to override.  And also update GbHeaderToolbar to take a model with the data it needs (rather than hitting the business service directly
