--- gradebookng/tool/src/java/org/sakaiproject/gradebookng/tool/component/GbDataTable.html
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<wicket:panel xmlns:wicket="http://wicket.apache.org">
-	<caption wicket:id="caption"></caption>
-	<thead wicket:id="topToolbars">
-	<wicket:container wicket:id="toolbars"></wicket:container>
-	</thead>
-	<tfoot wicket:id="bottomToolbars">
-	<wicket:container wicket:id="toolbars"></wicket:container>
-	</tfoot>
-	<tbody wicket:id="body">
-	<tr wicket:id="rows" role="row">
-		<td wicket:id="cells" tabindex="0">
-			<div wicket:id="cell">[cell]</div>
-		</td>
-	</tr>
-	</tbody>
-</wicket:panel>
